Api organizations

+++ app/tests/utils/mixins.py
@@ -7,7 +7,7 @@ from flask_fixtures import FixturesMixin
 from server import server
 from covidbed.model.abc import db
 
-server.config['TESTING'] = True
+server.config["TESTING"] = True
 
 
 class BaseTest(unittest.TestCase, FixturesMixin):
@@ -23,49 +23,32 @@ class BaseTest(unittest.TestCase, FixturesMixin):
 
 
 class BaseAuthMixin(object):
-
     def authenticate(self, email, password):
         response = self.client.post(
-            '/api/auth/login',
-            json={'email': email, 'password': password},
-            content_type="application/json"
-
+            "/api/auth/login", json={"email": email, "password": password}, content_type="application/json"
         )
 
         if response.status_code != 200:
             raise Exception("Invalide user")
-        result = json.loads(response.data.decode('utf-8'))
+        result = json.loads(response.data.decode("utf-8"))
         return result["token"]
 
     def get(self, url, token, params=None):
 
         return self.client.get(
-            url,
-            data=params,
-            headers={'Authorization': 'Bearer ' + token}
+            url, json=params, headers={"Authorization": "Bearer " + token}, content_type="application/json"
         )
 
     def post(self, url, token, params=None):
 
         return self.client.post(
-            url,
-            json=params,
-            headers={'Authorization': 'Bearer ' + token},
-            content_type='application/json'
+            url, json=params, headers={"Authorization": "Bearer " + token}, content_type="application/json"
         )
 
     def put(self, url, token, params=None):
         return self.client.put(
-            url,
-            json=params,
-            headers={'Authorization': 'Bearer ' + token},
-            content_type='application/json'
+            url, json=params, headers={"Authorization": "Bearer " + token}, content_type="application/json"
         )
 
     def delete(self, url, token):
-        return self.client.delete(
-            url,
-            headers={'Authorization': 'Bearer ' + token}
-        )
-
-
+        return self.client.delete(url, headers={"Authorization": "Bearer " + token})
